+++ ReleaseNotes.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+- Parametrized QT version
+- New distance units (scale widget)
+
 ## v1.0.4
 
 - Qt6 support